fuck

+++ java_package_import_list.py
@@ -51,8 +51,9 @@ class JavaPackageImportRef:
 
 
     def start(self, package_root_dir, used_dir):
-        os.remove(self.__result_file)
-		self.__total_used_files = 0
+        if os.path.exists(self.__result_file):
+            os.remove(self.__result_file)
+        self.__total_used_files = 0
         for package_dir, dirs, files in os.walk(package_root_dir, topdown=False):
             for name in dirs:
                 dir_path = os.path.join(package_dir, name)
@@ -65,5 +66,5 @@ class JavaPackageImportRef:
 
 if __name__ == '__main__':
     find_process = JavaPackageImportRef()
-    find_process.start('./moduleA/src', './src')
+    find_process.start('../base/src', './src')
     print('start....')
